requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache
redis==5.0.1

# Authentication and Security
passlib==1.7.4
python-jose==3.3.0

# Templates and Static Files
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6

# Validation and Serialization
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# HTTP Client
httpx==0.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Monitoring
loguru==0.7.2

# Environment and Configuration
python-dotenv==1.0.0